fix: address minimum-Git review findings
